feat: selectable AI provider, chat UX overhaul, phpstan level 8
Provider abstraction. OpenAIService hardcoded the completions URL at both
call sites and read its model from constructor config, so the app could
only talk to OpenAI with one model. Provider and model now travel with
the request: an AiProvider enum in Domain is the allowlist (OpenAI,
OpenRouter, Groq — all OpenAI-compatible, so the existing streaming code
works unchanged), and ModelSelection carries provider, model and an
optional user key for a single request. Absent selection reproduces
today's behaviour exactly.

A request never supplies a URL; the submitted provider is matched against
the enum and the URL always comes from the registry, closing the SSRF
this would otherwise open. User keys are marked SensitiveParameter, held
in a private property json_encode skips, kept out of log context, matched
against a strict character class that also blocks CRLF header injection,
and added to dontFlash. Tests assert a key reaches neither logs, exception
messages, nor the object's serialised forms.

Chat UX. The client and server disagreed about what a given input was:
the UI used "shorter than 10 characters or starts with 00000000" while
the backend used is_numeric() or the 64-hex-with-leading-zeros pattern.
"deadbeef" rendered as "Block #deadbeef" while the backend fetched it as
a transaction. The rule now lives in one place and is mirrored exactly;
all nine probe inputs agree, and a feature test pins the labels.

Streaming failures used to leave a skeleton pulsing forever. Both reader
loops are replaced by one consumeStream() that reports done, error,
stopped or interrupted, and each renders an actionable card: quota
exhaustion says when it resets, an unknown txid reads as an ordinary
empty result rather than a failure, and a truncated answer keeps the
partial text. That rewrite also fixed real token loss — the old loops
split each network read on newlines and silently dropped any SSE frame
straddling a read boundary.

Also: a stop button backed by AbortController, a skeleton that reflects
real progress, three one-click examples in place of an empty field, a
composer that stays one row on small screens, and nine icons that were
referenced in views but never registered, so they had been rendering as
nothing.

PHPStan level 5 to 8, one level at a time, no new ignores. Level 8 found
genuine null dereferences rather than annotation gaps: two Chat accessors
declared non-nullable returns that could be null, isBlock() fataled on an
empty chat, nullable message meta was passed into value objects that
reject null, and a webhook missing its Svix header was a 500 rather than
the 401 the handler already implements.

HistoryService called through() on a paginator and discarded the result,
relying on undocumented in-place mutation. Its test stubbed through() to
return null, which the real implementation never does, so the test only
passed because production threw the value away.

Level 9 is achievable but deliberately out of scope: 60 of its 70 errors
trace to $response->json() returning mixed across 23 files, and fixing it
properly means validating each external payload at the HTTP boundary.

+++ tests/Unit/Blockchain/PriceServiceTest.php
@@ -35,7 +35,7 @@ final class PriceServiceTest extends TestCase
         $http = self::createStub(HttpClientInterface::class);
         $logger = self::createStub(LoggerInterface::class);
         $now = Carbon::parse('2025-01-15');
-        $timestamp = Carbon::parse('2025-01-10')->timestamp;
+        $timestamp = (int) Carbon::parse('2025-01-10')->timestamp;
 
         $cache = $this->createMock(Repository::class);
         $cache->expects($this->once())
@@ -60,7 +60,7 @@ final class PriceServiceTest extends TestCase
         $cache = self::createStub(Repository::class);
         $logger = self::createStub(LoggerInterface::class);
         $now = Carbon::parse('2025-01-15');
-        $timestamp = Carbon::parse('2023-01-01')->timestamp; // More than 1 year ago
+        $timestamp = (int) Carbon::parse('2023-01-01')->timestamp; // More than 1 year ago
 
         $service = new PriceService($http, $logger, $cache, enabled: true, now: $now);
 
@@ -95,7 +95,7 @@ final class PriceServiceTest extends TestCase
     public function test_historical_price_fetches_from_api_when_not_cached(): void
     {
         $now = Carbon::parse('2025-01-15');
-        $timestamp = Carbon::parse('2025-01-10')->timestamp;
+        $timestamp = (int) Carbon::parse('2025-01-10')->timestamp;
 
         $response = $this->createMock(Response::class);
         $response->method('successful')->willReturn(true);
